# src/document/swagger-final.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Sekolah Kopi Raisa",
    "version": "1.0.0",
    "description": "API untuk aplikasi Sekolah Kopi Raisa - E-commerce dan Management System"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    },
    {
      "url": "https://be-web-sekolah-kopi-raisa.vercel.app",
      "description": "Vercel Production Server"
    },
    {
      "url": "https://sekolah-kopi-raisa.up.railway.app",
      "description": "Railway Production Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    }
  },
  "paths": {
...
    "/api/v1/news": {
      "get": {
        "tags": [
          "News"
        ],
        "summary": "Get all news",
        "responses": {
          "200": {
            "description": "News retrieved successfully"
          }
        }
      },
      "post": {
        "tags": [
          "News"
        ],
        "summary": "Create new news",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "content"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Judul berita (maksimal 90 karakter)",
                    "maxLength": 90,
                    "example": "Berita Terbaru Sekolah Kopi"
                  },
                  "content": {
                    "type": "string",
                    "description": "Konten berita (maksimal 2110 karakter)",
                    "maxLength": 2110,
                    "example": "Ini adalah konten berita yang informatif..."
                  },
                  "media": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Media files (maksimal 4 files)"
                  },
                  "thumbnail": {
                    "type": "string",
                    "format": "binary",
                    "description": "Thumbnail image"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "News created successfully"
          }
        }
      }
    },
    "/api/v1/news/{id}": {
      "get": {
        "tags": [
          "News"
        ],
        "summary": "Get news by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "News retrieved successfully"
          }
        }
      },
      "put": {
        "tags": [
          "News"
        ],
        "summary": "Update news",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "media": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "thumbnail": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "News updated successfully"
          }
        }
      },
      "delete": {
        "tags": [
          "News"
        ],
        "summary": "Delete news",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "News deleted successfully"
          }
        }
      }
    },
...
  }
}